requirements.txt
# Core Dependencies
PyQt5>=5.15.6
apscheduler>=3.9.1
pyyaml>=6.0
requests>=2.28.1
pillow>=9.3.0
python-dotenv>=0.21.0

# Video Processing
moviepy>=1.0.3
opencv-python>=4.6.0
numpy>=1.23.4

# Image Generation
stability-sdk>=0.5.0  # For Stable Diffusion
openai>=0.27.0  # For DALL-E

# Audio Processing
pydub>=0.25.1
gtts>=2.3.0  # Google Text-to-Speech

# Google APIs
google-api-python-client>=2.70.0
google-auth>=2.15.0
google-auth-oauthlib>=0.7.1
google-auth-httplib2>=0.1.0
gspread>=5.7.2

# Data Handling
pandas>=1.5.1

# Logging and Monitoring
rich>=12.6.0  # Better console output

# Testing
pytest>=7.2.0
pytest-mock>=3.10.0

# Development Tools
black>=22.10.0  # Code formatting
flake8>=6.0.0  # Linting
pre-commit>=2.20.0  # Pre-commit hooks